Create `uploadImageAndGetURL` util file
Related to #58
